+++ src/main/java/com/dgsaltarin/backend/domain/dto/CreatedAccountResponse.java
@@ -4,5 +4,5 @@ public class CreatedAccountResponse {
 
     private String userName;
     private double initialBalance;
-    private Long accountNumber;
+    private int accountNumber;
 }
